ask_hr_rag/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anyio==4.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
python-dotenv==1.2.1
google-genai==1.55.0
Jinja2==3.1.6
